+++ src/Word.php
@@ -1,23 +1,8 @@
 <?php
-/**
- * Wordsmith
- *
- * @author     Michael Rog <michael@michaelrog.com>
- * @link       https://topshelfcraft.com
- * @copyright  Copyright 2020, Top Shelf Craft (Michael Rog)
- * @see        https://github.com/topshelfcraft/Wordsmith
- */
-
-namespace topshelfcraft\wordsmith;
+namespace TopShelfCraft\Wordsmith;
 
 use Stringy\Stringy;
 
-
-/**
- * @author Michael Rog <michael@michaelrog.com>
- * @package Wordsmith
- * @since 3.0.0
- */
 class Word extends Stringy
 {
 }