+++ src/libs/FullNameParser.php
@@ -1,15 +1,5 @@
 <?php
-/**
- * Wordsmith
- *
- * @author     Michael Rog <michael@michaelrog.com>
- * @link       https://topshelfcraft.com
- * @copyright  Copyright 2020, Top Shelf Craft (Michael Rog)
- * @see        https://github.com/topshelfcraft/Wordsmith
- */
-
-namespace topshelfcraft\wordsmith\libs;
-
+namespace TopShelfCraft\Wordsmith\libs;
 
 /**
  * FullNameParser